+++ app/code/Svea/Maksuturva/Block/Form/Button.php
@@ -12,7 +12,7 @@ class Button extends \Magento\Config\Block\System\Config\Form\Field
             $html = $this->getLayout()->createBlock('Magento\Backend\Block\Widget\Button')
                 ->setData(
                     [
-                        'label' => __('Manually Query Maksuturva API For Orders Created In Past 3 Hours'),
+                        'label' => __('Manually Query Svea Payments API For Orders Created In Past 3 Hours'),
                         'onclick' => "setLocation('$url')",
                         'class' => 'action-add',
                     ]
@@ -22,7 +22,7 @@ class Button extends \Magento\Config\Block\System\Config\Form\Field
             $html = $this->getLayout()->createBlock('Magento\Backend\Block\Widget\Button')
                 ->setData(
                     [
-                        'label' => __('Manually Query Maksuturva API For Orders Created In Past Week'),
+                        'label' => __('Manually Query Svea Payments API For Orders Created In Past Week'),
                         'onclick' => "setLocation('$url')",
                         'class' => 'action-add',
                     ]