+++ app/code/Svea/Maksuturva/Block/Redirect/Maksuturva.php
@@ -33,7 +33,7 @@ class Maksuturva extends \Magento\Payment\Block\Form
         }
 
         $html = '<html><head></head><body>';
-        $html .= __('You will be redirected to the Maksuturva website in a few seconds.');
+        $html .= __('You will be redirected to the Svea Payments website in a few seconds.');
         $html .= $form->toHtml();
         $html .= '<script type="text/javascript">document.getElementById("maksuturva_payment_checkout").submit();</script>';
         $html .= '</body></html>';