+++ app/code/Svea/Maksuturva/Model/PaymentAbstract.php
@@ -186,8 +186,8 @@ abstract class PaymentAbstract extends \Magento\Payment\Model\Method\AbstractMet
             $this->_methods = $this->getGatewayImplementation()->getPaymentMethods($quoteTotal);
             if ($this->_methods) {
                 $this->cache->save(serialize($this->_methods), $cacheKey, array("MAKSUTURVA"), 60 * 5);
-                //
-                if ($this->_methods && count($this->_methods) > 0){
+                // added instanceof check as is_countable is only available PHP >7.3
+                if ($this->_methods && (is_array($this->_methods) || $this->_methods instanceof Countable) && count($this->_methods) > 0){
                     foreach($this->_methods as $method){
                         $this->_methodResourceModel->insert($method);
                     }